DataGridLite
Had to start a very light-weight data grid for a WPF project at work
running on some very, VERY low-powered machines. This is basically a
super-striped-down data grid with a very minimal number of features and
the fewest number of actual Visual elements I could get away with.

This is still in progress